Fixed the navbar by using a fixedTop prop on it
