+++ .gitignore
@@ -5,3 +5,5 @@
 node_modules/
 
 config.js
+
+project!-c26e6a0b17d4.json
