+++ public/index.html
@@ -1,12 +1,10 @@
 <html>
 <head>
 	<title>LandMarks!</title>
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap-theme.min.css">
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
 	<script src="node_modules/jquery/dist/jquery.js"></script>
 	<script src="node_modules/lodash/lodash.js"></script>
-	<link rel="stylesheet" href="https://cdn.rawgit.com/twbs/bootstrap/48938155eb24b4ccdde09426066869504c6dab3c/dist/css/bootstrap.min.css">
-	<script src="https://unpkg.com/reactstrap/dist/reactstrap.min.js"></script>
-
 </head>
 <body>
 <div class = "container"></div>
